Add installation instructions for Mac OS X
Added installation instructions for MacOS for style50. I decided to use HomeBrew for the installation because of the PyPa warning against installing PIP on the OS's distribution of python and because the easiest way to install libmagic and astyle is through homebrew. 

If we were to use the OS instance of python we would need to either provide more complicated instructions to install libmagic and astyle OR provide instructions on creating a symlink to the HomeBrew installation directory (I believe).